- Decorators have been split into two based on sync/async application - Type annotation check should now complain less when using those - New root files sync/async.py added to allow better control over imports
