+++ py_cachify/async.py
@@ -1,6 +1,6 @@
-from . import AsyncClient
-from .cached import async_cached
-from .lock import async_lock, async_once
+from .backend.cached import async_cached
+from .backend.lock import async_lock, async_once
+from .backend.types import AsyncClient
 
 
 __all__ = ['async_once', 'async_lock', 'async_cached', 'AsyncClient']
