+++ py_cachify/backend/exceptions.py
@@ -3,3 +3,7 @@ from __future__ import annotations
 
 class CachifyInitError(Exception):
     pass
+
+
+class CachifyLockError(Exception):
+    pass
