+++ py_cachify/backend/lock.py
@@ -6,9 +6,9 @@ from contextlib import asynccontextmanager, contextmanager
 from functools import partial, wraps
 from typing import Any, AsyncGenerator, Awaitable, Callable, Generator, ParamSpec, TypeVar, Union, cast
 
-from .backend.lib import get_cachify
 from .base import get_full_key_from_signature, is_coroutine
 from .exceptions import CachifyLockError
+from .lib import get_cachify
 
 
 logger = logging.getLogger(__name__)
