+++ py_cachify/sync.py
@@ -1,6 +1,6 @@
-from . import SyncClient
-from .cached import sync_cached
-from .lock import lock, sync_once
+from .backend.cached import sync_cached
+from .backend.lock import lock, sync_once
+from .backend.types import SyncClient
 
 
 __all__ = ['sync_once', 'sync_cached', 'lock', 'SyncClient']
