+++ tests/test_locks.py
@@ -1,6 +1,6 @@
 import pytest
 
-from py_cachify.exceptions import CachifyLockError
+from py_cachify import CachifyLockError
 from py_cachify.lock import async_lock, lock
