+++ tests/test_once_decorator.py
@@ -4,7 +4,7 @@ from time import sleep
 
 import pytest
 
-from py_cachify.exceptions import CachifyLockError
+from py_cachify import CachifyLockError
 from py_cachify.lock import once
